backend/major-service/requirements.txt
# 专业推荐API依赖包
fastapi==0.109.2
uvicorn[standard]==0.27.1
psycopg2-binary==2.9.9
pydantic==2.5.3
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0